backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy-orm
pydantic
langgraph
langchain
langchain-openai
python-dotenv
